+++ apps/frontend/src/types/maintenanceState.ts
@@ -1,0 +1,10 @@
+type Maintenance = {
+  id: number;
+  name: string;
+  label: string;
+  isInMaintenance: boolean;
+};
+
+export type MaintenanceStates = {
+  [id: string]: Maintenance;
+};
